renamed mrjob_spark_harness.py to simply harness.py
